input/ts/sec_ts: Fix up expected_device_id logic
sec_ts driver could not make up its mind whether the device ID
matched or not and reading 3 bytes from DT properly was too much
for this poor guy. Fix it up to let go of some spam.

+++ drivers/input/touchscreen/sec_ts/include/sec_ts.h
@@ -893,7 +893,7 @@ struct sec_ts_plat_data {
 	u8 config_version_of_bin[4];
 	u8 img_version_of_ic[4];
 	u8 img_version_of_bin[4];
-	u32 expected_device_id[3];
+	u8 expected_device_id[3];
 
 	struct pinctrl *pinctrl;
